ProjetoExemplo/src/test/java/br/com/fasam/projetoexemplo/CategoriaTest.java
/*
 * To change this template, choose Tools | Templates
 * and open the template in the editor.
 */
package br.com.fasam.projetoexemplo;

import br.com.fasam.projetoexemplo.entidades.Artigo;
import br.com.fasam.projetoexemplo.entidades.Categoria;
import br.com.fasam.projetoexemplo.entidades.Usuario;
import static junit.framework.Assert.assertNotNull;
import junit.framework.TestCase;

/**
 *
 * @author Aluno
 */
public class CategoriaTest extends TestCase {
    
    public CategoriaTest(String testName) {
        super(testName);
    }
    
    @Override
    protected void setUp() throws Exception {
        super.setUp();
    }
    
    @Override
    protected void tearDown() throws Exception {
        super.tearDown();
    }
    // TODO add test methods here. The name must begin with 'test'. For example:
    // public void testHello() {}
     public void testNomeEmCategoria(){
       Categoria categoria = new Categoria();
       assertNotNull(categoria.getNome());
   }
}
